refactor(visualization): split generators.py into focused submodules
generators.py was 1896 LOC, dominated by a 920-line _TEMPLATE r-string
(the full HTML/Jinja2 report) and a cluster of Mermaid sequence-diagram
helpers wedged between imports and the data layer.

Extracted into:
- visualization/_template.py — _TEMPLATE constant (928 LOC, mostly
  HTML/CSS/JS for the report)
- visualization/_mermaid.py — _mermaid_from_steps, _mermaid_trace,
  _safe_mermaid, _strip_markdown (96 LOC)

generators.py is now 896 LOC: data extraction, KPI computation, diff
rows, behavior summaries, generate_visual_report orchestration, and
_render_template. The remaining file is one cohesive responsibility:
turn EvaluationResult lists into the template render context.

Public API unchanged — `from evalview.visualization import
generate_visual_report` works as before.

Zero functional change. 1792 tests pass.
